66- silme- işlemlerinde -silinecek -öğe- bulunamadığında- exception -mı -olmamalı- bunun -yerine- başarılı -mesajı -dönülebilir

+++ src/main/java/tr/edu/ogu/ceng/service/FacultySupervisorService.java
@@ -109,16 +109,16 @@ public class FacultySupervisorService {
 
 	public boolean deleteFacultySupervisor(long id) {
 		try {
-			facultySupervisorRepository.deleteById(id);
-			log.info("Faculty supervisor deleted with id: {}", id);
-			return true;
+			if(facultySupervisorRepository.existsById(id)) {
+				facultySupervisorRepository.deleteById(id);
+				log.info("Faculty supervisor deleted with id: {}", id);
+			}		
 		} catch (DataIntegrityViolationException e) {
 			log.warn("Cannot delete faculty supervisor with ID {} due to integrity violation", id);
-			return false;
 		} catch (EmptyResultDataAccessException e) {
 			log.warn("Faculty supervisor with ID {} not found", id);
-			return false;
 		}
+		return true;
 	}
 
 	public Page<FacultySupervisorResponseDto> getAllFacultySupervisors(Pageable pageable) {
